refactor: modify contract calls support validation (#431)
* Update the transaction signer with stacks transaction
signing functionality. Add a TODO to one of the functions
* Replace the generic contract call wrapper struct
with an enum. Use the enum to send stacks sign messages
between the signers.
* Add the validate function to the contract trait.
These functions validate that the data in the trait
matches the data in the database.
* Change query to sort recent rotate key transactions
by their txid if more than one happened within a
bitcoin block span.