1444 chorebug restrict wipe databases to testing api (#1445)
## Description

Closes #1444 

The endpoint was already disabled in the private API (used by the
sidecar) and will now only be accessible through the testing API and
client.

I'm not very happy with the `as_testing()` method and am open to
suggestions. Ideally, all test calls to Emily should go through the
test-emily-client, but this would require a larger refactor since the
structs used in the two crates are different.

## Changes
- disable the wipe_databases endoint.
- add the test-emily-client crate to the dev configuration of the
signer.
- use the test-emily-client crate to call the wipe-databases.  


- [x] I have performed a self-review of my code
- [x] My changes generate no new warnings
- [x] New and existing unit tests pass locally with my changes
- [x] Any dependent changes have been merged and published in downstream
modules