chore: add comments for `get_swept_*` queries near context boundary (#1482)
## Description

Closes: #?

## Changes

The `get_swept_deposit_requests` and `get_swept_withdrawal_requests`
queries may return completed requests as swept and pending if the
relative completed events are anchored to a bitcoin block that precede
the sweep bitcoin block.
This is fine as we have other layers taking care of this:
- The signers validating the contract call will ask the stacks node it
the request is still pending
 - The contracts rejecting replay

This PR add comments about it.

## Testing Information

Add a test to show this situation.

## Checklist:

- [ ] I have performed a self-review of my code
- [ ] My changes generate no new warnings
- [ ] New and existing unit tests pass locally with my changes
- [ ] Any dependent changes have been merged and published in downstream
modules