chore: emily remove private endpoints (#1474)
## Description

Remove the declaration of a couple of endpoints that are not meant to be
called by the signers from the public API.

## Changes
- remove `create_withdrawal` from the public API. The withdrawals are
expected to be created from the new-block endpoint. We keep it in the
**private** API in order to have a way to recreate missing withdrawals
in an easy manner in case of a problem.

- remove `set_chainstate` from the public API. Only the sidecar is
supposed the endpoint throught the private API.

## Checklist:

- [x] I have performed a self-review of my code
- [x] My changes generate no new warnings
- [x] New and existing unit tests pass locally with my changes
- [x] Any dependent changes have been merged and published in downstream
modules