fix: stop coordinator activities when tenure finishes (#1518)
## Description

Closes https://github.com/stacks-network/sbtc/issues/1516.

I thought about adding just the bitcoin chain tip block height to the
signer state, since that is all that is needed for this PR; plus an
atomic integer is nicer. But a `BitcoinBlockRef` is probably be more
useful, generally speaking, so here we are.

## Changes

* Add the bitcoin chain tip into the signer state object. Set it in the
block observer when a new block is observed.
* End the coordinator's tenure when creating deposit and withdrawal
stacks transactions if the current chain tip doesn't match the one that
was used at the start of the coordinator's tenure.

## Testing Information

I tested that this solves the problem described in
https://github.com/stacks-network/sbtc/issues/1516 in devenv, and it
works as it's supposed to.

## Checklist:

- [x] I have performed a self-review of my code