feat: add encoder/decoder for withdrawal id compression (for storage in OP_RETURN) (#1525)
## Description

Part of: #1478 and supersedes #1486 (due to the size of the changes and
the fact that most comments are now invalidated).

This introduces a new module `idpack` to the `sbtc` crate which contains
an encapsulated integer sequence compression encoding using a
multi-bitmap segmentation algorithm.

Compared to #1486, this PR is substantially slimmed down and drops
support for the eventual handling of multiple encodings within the same
encoding format, i.e. adding or changing the encodings will require a
new implementation.

# Changes

- Removed `EncodingStrategy` trait and moved the full encode/decode
implementations to `encoder.rs` and `decoder.rs` respectively.
- Removed a bunch of other stuff as well.
- The LEB128 implementation is moved to #1523 and this PR targets that.
- The encode-path is now infallible. 
- The packaging path remains fallible:
- The only errors that can be produced (in practice) are due to input
validation.
  - The same applies to size-estimations from `Segmenter`.
- Reason being that these work on slices instead of requiring new allocs
to sortable types (we expect our input to be sorted, anyway - see [this
comment](https://github.com/stacks-network/sbtc/pull/1512#discussion_r1994255206)
for more context).
- Added a bunch of tests.
- The `Segments` type now upholds the invariant that its inner segments
must be sequential without overlaps.

## Checklist:

- [x] I have performed a self-review of my code
- [x] My changes generate no new warnings
- [x] New and existing unit tests pass locally with my changes
- [x] Any dependent changes have been merged and published in downstream
modules