chore: bump rust version to 1.85.1 (#1552)
## Description

Bump the version of Rust to 1.85.1. The diff is due to sorting
differences.

## Changes

* Bump the rust version in the workspace. Also update the base images
used in our rust containers.
* Use the 2024 edition in the `signer`, `sbtc`, `emily`,
`blocklist-client`, and `blocklist-openapi-gen` crates. The other
generated crates remain unchanged.
* The `std::env::set_var` function and `std::env::remove_var` functions
are now unsafe, so made some changes related to that in our tests.
* A few changes in the signer crate related to new things that clippy
picks up.
* Everything else is related to how rust now sorts imports in this
edition. It is explained in
https://doc.rust-lang.org/edition-guide/rust-2024/rustfmt-version-sorting.html.
This is where all the noise diffs come from.

## Testing Information

No functionality has changed, so if tests pass then we should be good.

## Checklist:

- [x] I have performed a self-review of my code

---------

Co-authored-by: cylewitruk <cyle.witruk@outlook.com>