feat: signers no longer store transactions in the database (#1559)
## Description

Closes https://github.com/stacks-network/sbtc/issues/1557 and closes
https://github.com/stacks-network/sbtc/issues/1466.

## Changes

* Drop the `transactions.tx` column in the signer database.
* Remove the `tx` field from the `model::Transaction` struct.
* Remove the `DbRead::get_bitcoin_tx` method, since we cannot implement
it with the `transactions.tx` column removed. Also, it was unused.
* Remove the `BitcoinTx` struct.
* Ignore the in-memory version of the signer UTXO tests. This is because
the `TestData` struct does not correctly populate the database with
`TxPrevouts` or `TxOutputs`. Might be worth fixing at some later point.

## Testing Information

This is kind of a refactor. No new tests, but if the existing tests pass
then we should be good. Granted, this PR ignores some tests...

## Checklist:

- [x] I have performed a self-review of my code