feat: introduce p2p connection limiting (`#1547`) (#1568)
## Description

Closes: #1547

## Changes
- Makes use of the built-in libp2p `ConnectionLimits` behavior to
implement some (hopefully) sane limits on connections.
- Adds a few utility functions used in tests
- Adds enable/disable for both kademlia and autonat (these both affect
connections, so needed to disable them for testing).
- Adds a context event for the p2p event loop having started (also used
in test)

## Testing Information

This proved to be a bit trickier to test than originally thought. I
added a test for the "big one" which is total established connections,
but we could probably add a few more to test the other limits (I have
seen some of the other errors while fiddling with tests, though).
Otherwise, this is tested in `devenv` and things seem to work just fine
there, but it would be nice to test on `devnet` to ensure that the
limits are indeed sane.

## Checklist:

- [x] I have performed a self-review of my code
- [x] My changes generate no new warnings
- [x] New and existing unit tests pass locally with my changes
- [x] Any dependent changes have been merged and published in downstream
modules